moved batchManager
