--- BusinessLayer/BatchModule/IBatchManager.cs
@@ -1,17 +1,0 @@
-﻿using BeerProductionSystem.DOClasses;
-using System;
-
-namespace BeerProductionSystem.BusinessLayer.BatchModule
-{
-    interface IBatchManager
-    {
-        BatchDO CurrentBatch { get; set; }
-        StateLogDO StateLog { get; set; }
-
-        void CreateBatch(ushort productType, ushort productionSpeed, ushort batchSize);
-
-        bool CheckBatchParameter();
-
-        void SaveTimeInState(MachineState currentState, TimeSpan timeSpan);
-    }
-}
